Add DataConfig
Represents the various configurable properties of the game elements.

The not so apparent configurable properties are:

1. RowInset and ColumnInset in PelletConfig: A Grid with 4 rows and 4
columns and value 1 to both RowInset and ColumInset would leave the
center 2x2 grid for pellets to become active at, as the insets would
work as the border to the grid.

2. TimeSkipPerFrame: This represents how quickly the snake moves. The
value is in seconds.