Add Snake
The Initialize method creates a snake with some initial size.

The GetNextCell returns the next cell that the snake’s head would be at
when it follows the next direction.

The AddHead method adds the next cell as the snake’s head.

The RemoveTail method removes the last cell from the snake’s body.

The Deactivate method deactivates the whole snake.