Add BindInputHandlerCommand
This command binds the right input handler with to the IInputHandler
interface that can then be injected where its required.

If the player wants to control the snake itself, either the
KeyboardHandler or the SwipeHandler is bound, based on the device.

If its the AI, then AIHandler is bound to the IInputHandler along with
the right kind of pathfinder.